Cleanup #11 - some code cleanup and organizing
- removed test text files
- updated .gitignore to block python cache
- renamed project directory
- renamed main py file
- created main executible
- retructured code for remote execution to avoid race conditions
