+++ .gitignore
@@ -1,0 +1,2 @@
+.mypy_cache
+__pycache__
