--- Second-file.dgt
@@ -1,1 +1,0 @@
-I got save working.
