--- first-note.dgt
@@ -1,1 +1,0 @@
-This is my first file from writing my very own text edit app.
